+++ eds_pseudo/__init__.py
@@ -1,4 +1,4 @@
 from pathlib import Path
-import eds_pseudonymisation.pipelines
+import eds_pseudo.pipes
 
 BASE_DIR = Path(__file__).parent