+++ eds_pseudo/pipes/dates/patterns.py
@@ -1,5 +1,5 @@
-from edsnlp.pipelines.misc.dates.patterns.absolute import absolute_pattern, mode_pattern
-from edsnlp.pipelines.misc.dates.patterns.atomic.delimiters import raw_delimiter_pattern
+from edsnlp.pipes.misc.dates.patterns.absolute import absolute_pattern, mode_pattern
+from edsnlp.pipes.misc.dates.patterns.atomic.delimiters import raw_delimiter_pattern
 from edsnlp.utils.regex import make_pattern
 
 numeric_day_pattern = r"(?<!\d)(0 [1-9]|[12] \d|3 [01])(?!\d)"